# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is based on the Philippine Standard Geographic Code (PSGC), which is a systematic classification and coding of geographic areas in the Philippines. Its units of classification are based on the four well-established levels of geographical-political subdivisions of the country, including the region, the province, the municipality/city, and the barangay.",
        "title": "Philippine Standard Geographic Code (PSGC) API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/municipalities": {
            "get": {
                "description": "get Municipalities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Municipalities"
                ],
                "summary": "Show list of Municipalities",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedCityMuni"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/municipalities/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Municipalities"
                ],
                "summary": "Show a Municipality",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Municipality PsgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CityMuni"
                        }
                    },
                    "400": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/provinces": {
            "get": {
                "description": "get Provinces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provinces"
                ],
                "summary": "Show list of Provinces",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedProvince"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/provinces/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provinces"
                ],
                "summary": "Show a Province",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province PsgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Province"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/regions": {
            "get": {
                "description": "get Regions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Show list of Regions",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedRegion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/regions/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Show a Region",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region PsgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Region"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CityMuni": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prov_code": {
                    "type": "string"
                },
                "psgc_code": {
                    "type": "string"
                }
            }
        },
        "MetaData": {
            "type": "object",
            "properties": {
                "item_count": {
                    "type": "integer",
                    "example": 1000
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 1000
                },
                "total_items": {
                    "type": "integer",
                    "example": 10000
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "PaginatedCityMuni": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CityMuni"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/MetaData"
                }
            }
        },
        "PaginatedProvince": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Province"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/MetaData"
                }
            }
        },
        "PaginatedRegion": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Region"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/MetaData"
                }
            }
        },
        "Province": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "psgc_code": {
                    "type": "string"
                },
                "regCode": {
                    "type": "string"
                }
            }
        },
        "Region": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "psgc_code": {
                    "type": "string"
                }
            }
        }
    },
...
}